Reorganizing radar helpers
